Add files via upload
to a loss position with a value of MinProfitLock= -50 in the name of the deposit, we open the opposite pending position of Buy or Sell at a distance of Level Lock =25 pips with a double volume = Lot*2. If the losing position reduces the loss, the new pending position moves behind the losing position at the Level Lock distance. If the losing position increases its loss again, the pending order is stopped, waiting to be hit by the market. When this happens, an order opposite to the losing one is opened from the pending position in a double volume. If the market continues in this direction, the first position increases the loss and the new position takes a profit. When the profit and loss are equalized, we close both positions together, after reaching OrdersProfitLock= -1 in the name of the deposit. This is how we eliminate the loss of position. The parameter QtyLock=3 determines how many unprofitable positions we can liquidate at once in the same way, from the most unprofitable to the least unprofitable.


Can you help me with that?

Bronff


I don't have good experience with programming or github. I would like to find help. I do EA for mql4, mql5